refactor: Split the two suppression concerns into separate files
- Collecting the relevant data and
- Writing it in the specified format to files

The goal is to simplify the code for reading, especially the unit tests
